+++ kitty/utils.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 # License: GPL v3 Copyright: 2016, Kovid Goyal <kovid at kovidgoyal.net>
 
-import atexit
 import fcntl
 import math
 import os
@@ -36,7 +35,7 @@ from .constants import (
 from .fast_data_types import WINDOW_FULLSCREEN, WINDOW_MAXIMIZED, WINDOW_MINIMIZED, WINDOW_NORMAL, Color, Shlex, get_options, monotonic, open_tty
 from .fast_data_types import timed_debug_print as _timed_debug_print
 from .types import run_once
-from .typing import AddressFamily, PopenType, Socket, StartupCtx
+from .typing import AddressFamily, PopenType, StartupCtx
 
 if TYPE_CHECKING:
     import tarfile
@@ -350,17 +349,6 @@ class startup_notification_handler:
             end_startup_notification(self.ctx)
 
 
-def remove_socket_file(s: 'Socket', path: Optional[str] = None, is_dir: Optional[Callable[[str], None]] = None) -> None:
-    with suppress(OSError):
-        s.close()
-    if path:
-        with suppress(OSError):
-            if is_dir:
-                is_dir(path)
-            else:
-                os.unlink(path)
-
-
 def unix_socket_directories() -> Iterator[str]:
     import tempfile
     home = os.path.expanduser('~')
@@ -381,31 +369,6 @@ def unix_socket_paths(name: str, ext: str = '.lock') -> Generator[str, None, Non
     for loc in unix_socket_directories():
         filename = ('.' if loc == home else '') + name + ext
         yield os.path.join(loc, filename)
-
-
-def random_unix_socket() -> 'Socket':
-    import shutil
-    import socket
-    import stat
-    import tempfile
-
-    from kitty.fast_data_types import random_unix_socket as rus
-    try:
-        fd = rus()
-    except OSError:
-        for path in unix_socket_directories():
-            ans = socket.socket(family=socket.AF_UNIX, type=socket.SOCK_STREAM, proto=0)
-            tdir = tempfile.mkdtemp(prefix='.kitty-', dir=path)
-            atexit.register(remove_socket_file, ans, tdir, shutil.rmtree)
-            path = os.path.join(tdir, 's')
-            ans.bind(path)
-            os.chmod(path, stat.S_IRUSR | stat.S_IWUSR)
-            break
-    else:
-        ans = socket.socket(family=socket.AF_UNIX, type=socket.SOCK_STREAM, proto=0, fileno=fd)
-    ans.set_inheritable(False)
-    ans.setblocking(False)
-    return ans
 
 
 def parse_address_spec(spec: str) -> tuple[AddressFamily, Union[tuple[str, int], str], Optional[str]]: